+++ CHANGELOG.md
@@ -505,3 +505,4 @@
 - 2026-04-19: Wire mod, tz country, mod, geo, +3
 - 2026-04-20: Test send, conf, mod, site config, +5
 - 2026-04-21: Test summaries, mod, send, site config, +2
+- 2026-04-22: Test embeddings, mod, summaries, send, +1